GPSSensorTimeoutStrategy contains weird code (might be a merge issue) -> replaced with guessed interface\; DefaultServiceDictionary maps Services to Object -> created Service interface as supertype of Controller -> ServiceDictionary manages only Service instances (including Controller)\; renamed EventHander.getType to getTargetEventType\; If DefaultTrafficManager registers himself to the EventHandlerManager and VehicleEventHandlerManager, it has to be a EventHandler and a VehicleEventHandler itself -> stubs throwing UnsupportedOperationException added\; SimulationEventHandler is not an interface because it manages SimulationEvent which is simply an implementation of Event (and therefore should be named DefaultEvent) -> made SimulationEventHandler an implementation named AbstractEventHandler\; AbstractEventHandlerManager has no corresponding non-abstract Event type and therefore it's EJB implementation DefaultEventHandlerManager is unnecessary. VehicleEvent has to extends TrafficEvent in order to let TrafficServer handle both using the EventHandler API (this make sense and the generics forbid the current solution (with type unsafe cast) for a reason)\; moving TrafficEvent and TrafficEventHandler to traffic module\; making classes widely type safe\; removing passing of UUID id to Vehicle (should correspond to the common id pattern of using an immutable Long)

+++ simulation/de.pgalise.simulation.operationCenter/src/main/java/de/pgalise/simulation/operationCenter/internal/message/RemoveSensorsMessage.java
@@ -21,12 +21,12 @@ import java.util.Collection;
  * If you want to remove more than one sensor at the same time.
  * @author Timo
  */
-public class RemoveSensorsMessage extends OCWebSocketMessage<Collection<Long>> {
+public class RemoveSensorsMessage extends OCWebSocketMessage<Collection<Integer>> {
 	/**
 	 * Constructor
 	 * @param content a list with all sensor ids that will be removed.
 	 */
-	public RemoveSensorsMessage(Collection<Long> content) {
+	public RemoveSensorsMessage(Collection<Integer> content) {
 		super(OCWebSocketMessage.MessageType.REMOVE_SENSORS, content);
 	}
 }
